Get rid of lib/sapphire_bot/helpers
