+++ lib/sapphire_bot.rb
@@ -10,34 +10,86 @@ require 'discordrb'
 require 'terminal-table'
 require 'youtube-dl.rb'
 
-require_relative 'sapphire_bot/logger'
+# Methods that should be accessible everywhere.
+module Kernel
+  # Runs a block with warning messages supressed.
+  def run_supressed(&block)
+    original_verbosity = $VERBOSE
+    $VERBOSE = nil
+    yield block
+    $VERBOSE = original_verbosity
+  end
 
-# Supresses warning: already initialized constant Discordrb::LOGGER.
-original_verbosity = $VERBOSE
-$VERBOSE = nil
+  # Converts seconds to human readable format.
+  def time_in_words(time)
+    days = (time / 86_400).to_i
+    time -= days * 86_400
+    hours = (time / 3600).to_i
+    time -= hours * 3600
+    minutes = (time / 60).to_i
+    string = "#{days} day#{'s' unless days == 1},"
+    string << " #{hours} hour#{'s' unless hours == 1},"
+    string << " #{minutes} minute#{'s' unless minutes == 1}"
+  end
 
-# Set debug mode if command line arguments include "-debug".
-debug = ARGV.include?('-debug') ? true : false
-Discordrb::LOGGER = SapphireBot::LOGGER = if debug
-                                            SapphireBot::Logger.new(:debug)
-                                          else
-                                            SapphireBot::Logger.new
-                                          end
+  # Converts boolean values to more appealing format.
+  # Possible modes: on, enabled.
+  def bool_to_words(bool, mode = :on)
+    case mode
+    when :on
+      string_if_true = 'on'
+      string_if_false = 'off'
+    when :enabled
+      string_if_true = 'enabled'
+      string_if_false = 'disabled'
+    else
+      raise ArgumentError
+    end
 
-$VERBOSE = original_verbosity
+    return string_if_true if bool
+    string_if_false
+  end
 
-require_relative 'sapphire_bot/other/helpers'
-require_relative 'sapphire_bot/other/store_data'
-require_relative 'sapphire_bot/config'
-require_relative 'sapphire_bot/music_bot'
+  # Returns true if specified url string is valid.
+  def valid_url?(url)
+    uri = URI.parse(url)
+    return true if uri.is_a?(URI::HTTP) || uri.is_a?(URI::HTTPS)
+    false
+  rescue
+    false
+  end
 
-SapphireBot::CONFIG = SapphireBot::Config.new
+  # Returns urls host.
+  def url_host(url)
+    URI.parse(url).host
+  rescue
+    nil
+  end
+end
 
-Dir["#{File.dirname(__FILE__)}/sapphire_bot/*.rb"].each { |file| require file }
-
-require_relative 'discordrb/server'
-
+# Base module for sapphire.
 module SapphireBot
+  require_relative 'sapphire_bot/logger'
+
+  # Set debug mode if command line arguments include "-debug".
+  run_supressed do
+    debug = ARGV.include?('-debug') ? true : false
+    Discordrb::LOGGER = SapphireBot::LOGGER = if debug
+                                                SapphireBot::Logger.new(:debug)
+                                              else
+                                                SapphireBot::Logger.new
+                                              end
+  end
+
+  require_relative 'sapphire_bot/other/store_data'
+  require_relative 'sapphire_bot/config'
+  require_relative 'sapphire_bot/music_bot'
+
+  CONFIG = Config.new
+
+  Dir["#{File.dirname(__FILE__)}/sapphire_bot/*.rb"].each { |file| require file }
+
+  require_relative 'discordrb/server'
 
   BOT = Discordrb::Commands::CommandBot.new(token: CONFIG.discord_token,
                                             application_id: CONFIG.discord_client_id,
@@ -51,7 +103,7 @@ module SapphireBot
   Events.include!
 
   at_exit do
-    LOGGER.info "Saving files and deleting songs before exiting..."
+    LOGGER.info 'Saving files and deleting songs before exiting...'
     STATS.save
     ServerConfig.save
     MusicBot.delete_files
@@ -59,6 +111,6 @@ module SapphireBot
   end
 
   LOGGER.info "Oauth url: #{BOT.invite_url}+&permissions=#{CONFIG.permissions_code}"
-  LOGGER.info "Use ctrl+c to safely stop the bot."
+  LOGGER.info 'Use ctrl+c to safely stop the bot.'
   BOT.run
 end
