+++ lib/sapphire_bot/events/message.rb
@@ -3,7 +3,6 @@ module SapphireBot
     # Adds one to messages read stat every time a new message is received.
     module MessagesReadStat
       extend Discordrb::EventContainer
-      extend Helpers
       message(starting_with: not!(CONFIG.prefix),
               private: false) do |event|
         STATS.messages_read += 1 unless event.author.current_bot?
@@ -13,14 +12,13 @@ module SapphireBot
     # Automatically shortens urls.
     module AutoShorten
       extend Discordrb::EventContainer
-      extend Helpers
       message(starting_with: not!(CONFIG.prefix),
               private: false) do |event|
         unless event.from_bot?
           bot_profile = event.bot.profile.on(event.server)
           if bot_profile.permission?(:manage_messages, event.channel) &&
              event.server.shortening
-            text = shorten_text(event)
+            text = GOOGLE.shorten_text(event)
             unless event.message.content == text
               event.send_message("**#{event.author.username}**: #{text}")
               event.message.delete
