+++ lib/sapphire_bot/google_services.rb
@@ -2,7 +2,6 @@ module SapphireBot
   # Shortens links and finds videos from youtube.
   class GoogleServices
     include StoreData
-    include Helpers
 
     def initialize
       @youtube = Google::Apis::YoutubeV3::YouTubeService.new
@@ -24,6 +23,38 @@ module SapphireBot
       LOGGER.log_exception e
       nil
     end
+
+    # Detects urls in text and shortens them with goo.gl url shortener
+    # Can be called as shorten(event) (if event is Discordrb::Events::MessageEvent
+    #               or shorten(text, attributes)
+    def shorten_text(var, preview: true, original: true, minlength: 21)
+      if var.is_a?(Discordrb::Events::MessageEvent)
+        text = var.message.content
+        preview = var.server.preview if preview.nil?
+        original = var.server.original if original.nil?
+        minlength = var.server.minlength if minlength.nil?
+      elsif var.is_a?(String)
+        text = var
+      end
+
+      return text if !text || text.length < minlength
+
+      shortened_text = text.clone
+
+      URI.extract(text) do |url|
+        next if url.length < minlength
+        shortened_url = shorten_url(url)
+        unless shortened_url == url
+          shortened_url.insert(0, '<').insert(-1, '>') unless preview
+          shortened_url.insert(0, "(#{url_host(url)}) ") if original
+          shortened_text.gsub!(url, shortened_url)
+        end
+      end
+
+      shortened_text
+    end
+
+    private
 
     def shorten_url(url)
       if !@ignored_urls.any? { |ignored_url| url.include?(ignored_url) }
